dtc: Fix uninitialized value in local fixup generation
The local_fixup_generated member of struct fixup_entry was not
initialized on structure creation, causing fixups to randomly be left
out of the compiled overlay. This generally manifested itself with
kernel messages like

  could not find pctldev for node /ocp/interrupt-controller@48200000, deferring probe

as the kernel tried to follow a pinctrl phandle and instead got
whichever one happened to have the same ID as the overlay used
internally (before fixups).

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>